заменил float на decimal из-за  погрешности

+++ server/models/Product.js
@@ -6,7 +6,7 @@ const Product = sequelize.define('product', {
     id: {type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true},
     name: {type: DataTypes.STRING, allowNull: false},
     url: {type: DataTypes.STRING, defaultValue: null},
-    price: {type: DataTypes.FLOAT, allowNull: false},
+    price: {type: DataTypes.DECIMAL, allowNull: false},
     rating: {type: DataTypes.FLOAT, defaultValue: 0},
     img: {type: DataTypes.STRING(1024), allowNull: false},
     have: {type: DataTypes.INTEGER, defaultValue: 0},
